feat(#209): Login Response 이후 처리
