+++ src/pages/auth/RedirectPage/index.tsx
@@ -1,36 +1,38 @@
 import { useEffect } from 'react';
-import { useParams } from 'react-router-dom';
+import { useNavigate, useParams } from 'react-router-dom';
 
-import { get, post } from '@_utils/api';
+import { get } from '@_utils/api';
+
+interface LoginResponse {
+  status: 'ACTIVE' | 'INACTIVE';
+  access: string;
+  refresh: string;
+  access_expires_at: string;
+  refresh_expires_at: string;
+  user_data?: {
+    email: string;
+    nickname: string;
+  };
+}
 
 function RedirectPage() {
   const { type } = useParams();
+  const navigate = useNavigate();
 
   useEffect(() => {
     const code = new URL(document.location.toString()).searchParams.get('code');
-    console.log(code);
-    switch (type) {
-      case 'kakao':
-        get(`/user/kakao/callback/?code=${code}`).then((response) => {
-          console.log(response);
-        });
-        break;
 
-      case 'naver':
-        get(`/user/naver/callback/?code=${code}`).then((response) => {
-          console.log(response);
-        });
-        break;
+    get<LoginResponse>(`/user/${type}/callback/?code=${code}`).then((response) => {
+      console.log(response);
 
-      case 'google':
-        get(`/user/google/callback/?code=${code}`).then((response) => {
-          console.log(response);
-        });
-        break;
-
-      default:
-        break;
-    }
+      localStorage.setItem('access_token', response.data.access);
+      localStorage.setItem('refresh_token', response.data.refresh);
+      if (response.data.status === 'ACTIVE') {
+        navigate('/');
+      } else {
+        // navigate(`/signup?type=naver&email=${formdata.get('email')}&nickname=`);
+      }
+    });
   }, []);
 
   return <div>asdf</div>;
